APIWrapper and some cleanup
